fix: fix graph callbacks
