--- pages/__init__.py
@@ -1,6 +1,0 @@
-# flake8: noqa
-"""
-Pages package for Garmin Dashboard.
-"""
-
-from . import activity_detail, calendar, garmin_login
